pyaiot/gateway/ws: reorganize files and object names
