+++ pyaiot/gateway/ws/__init__.py
@@ -27,4 +27,4 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from .gateway import run
+from .application import run
